Fix deletion of nonexisting item, only delete if exists

+++ tests/installation.bats
@@ -17,29 +17,27 @@ function setup {
 }
 
 function teardown {
-    ddev rm -R ${SPRINT_NAME}
-    rm -rf ${SPRINTDIR}/${SPRINT_NAME}
+    ddev rm -R --omit-snapshot ${SPRINT_NAME}
+    if [ ! -z "${SPRINTDIR}" -a ! -z "${SPRINT_NAME}" -a -d ${SPRINTDIR}/${SPRINT_NAME} ] ; then
+        rm -rf ${SPRINTDIR}/${SPRINT_NAME};
+    fi
 }
 
 @test "check ddev project status and router status, check http status" {
 
     [ -f ${SPRINTDIR}/SPRINTUSER_README.md -a -f ${SPRINTDIR}/COPYING -a -d ${SPRINTDIR}/licenses ]
     DESCRIBE=$(cd ${SPRINTDIR}/${SPRINT_NAME} && ddev describe -j)
-    ROUTER_STATUS=$(echo $DESCRIBE | jq -r ".raw.router_status" )
+    ROUTER_STATUS=$(echo ${DESCRIBE} | jq -r ".raw.router_status" )
     [ "$ROUTER_STATUS" = "healthy" ]
 
-    STATUS=$(echo $DESCRIBE | jq -r ".raw.status")
+    STATUS=$(echo ${DESCRIBE} | jq -r ".raw.status")
     [ "$STATUS" = "running" ]
 
     DESCRIBE=$(cd ${SPRINTDIR}/${SPRINT_NAME} && ddev describe -j)
-    NAME=$(echo $DESCRIBE | jq -r ".raw.name")
-    HTTP_PORT=$(echo $DESCRIBE | jq -r ".raw.router_http_port")
+    NAME=$(echo ${DESCRIBE} | jq -r ".raw.name")
+    HTTP_PORT=$(echo ${DESCRIBE} | jq -r ".raw.router_http_port")
     URL="http://${DHOST}:${HTTP_PORT}"
     CURL="curl --fail -H 'Host: ${NAME}.ddev.local' --silent --output /dev/null --url $URL"
     echo "# curl: $CURL" >&3
-    $CURL
+    ${CURL}
 }
-
-
-# todo:
-# Test that drush works inside container
